+++ metrics-operator/controllers/common/providers/common.go
@@ -11,12 +11,3 @@ var SupportedProviders = []string{
 	PrometheusProviderType,
 	DataDogProviderType,
 }
-
-func IsProviderSupported(providerName string) bool {
-	for _, supportedProvider := range SupportedProviders {
-		if providerName == supportedProvider {
-			return true
-		}
-	}
-	return false
-}
